fix(logic): :bug: Bug fix
Fixed a bug that occurred when pressing the `Stop` button if the `wakapi` process was not running.

+++ src/logic.py
@@ -75,7 +75,8 @@ class Logic:
             sleep(self.timeout)
 
     def stop(self):
-        self.wakapi_process.kill()
+        if self.wakapi_process:
+            self.wakapi_process.kill()
 
 
 if __name__ == "__main__":
